feat: docker restart helper
because i was tired of going into the ssh machine and doing the command by hand. but actually... there is an issue where we need to restart the container for lobe-chat over and over again due to some mcp connection problem. now that we have to share the app to other users, it would be helpful to give them control to restart the chat interface / hard reset / aapne on off karke dekha?